Add files via upload
These are the .js files that will be used as models for the db, controllers to connect the html to the db and routes to direct the connections to the functions within the controller scripts. I made some edits to the login, profile, upload and sign up pages to allow the scripts to run and store the data in the db. I am currently working on the rental, review, chat, and payment